backend/requirements2.txt
fastapi
uvicorn
openai
langchain
langgraph
langchain-community
langchain-core
sqlalchemy
pymysql
python-dotenv
requests
charset_normalizer
pydantic<2.0
